+++ utils/graphql-pydantic-converter/CHANGELOG.md
@@ -15,5 +15,10 @@
 # 1.0.0
 - Migration to pydantic v2
 
-# 1.1.0
+# 1.0.1
 - Change Map type to dict (key, value)
+
+# 1.0.2
+- Stringify mutation input strings
+- Add __typename to payload
+- create test folder as a module